requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow==2.18.0
